Audio Playback Completed
Added in all the necessary code that is used for playback of the audio
sample. Changed a few of comments, made some modifications to the
interrupt handler. I also made it so that the PWM driver is only enabled
when the speaker is being used for playback to eliminate noise.